+++ surpyval/alpha/__init__.py
@@ -1,0 +1,9 @@
+"""Alpha-stage models.
+
+The least mature tier of SurPyval: interfaces here are exploratory and
+may change or disappear without notice. More mature (but still
+pre-stable) models live in ``surpyval.beta``.
+"""
+
+from .parallel import ParallelModel
+from .series import SeriesModel
