+++ surpyval/beta/__init__.py
@@ -1,0 +1,7 @@
+"""Beta-stage models.
+
+Models that are functionally complete and tested but whose interfaces
+have not yet stabilised into the release contract. The survival tree
+and random survival forest live in ``surpyval.beta.ml``. Less mature,
+exploratory models live in ``surpyval.alpha``.
+"""
