+++ surpyval/beta/ml/__init__.py
@@ -1,7 +1,7 @@
 """Machine-learning survival models.
 
 Home of SurPyval's tree-based models: the survival tree and the random
-survival forest. See ``surpyval.ml.forest`` for the implementation.
+survival forest. See ``surpyval.beta.ml.forest`` for the implementation.
 """
 
 from .forest import RandomSurvivalForest, SurvivalTree
