+++ surpyval/experimental/__init__.py
@@ -1,6 +1,20 @@
-# The survival tree and random survival forest graduated to
-# `surpyval.ml`; they are re-exported here for backwards compatibility.
-from surpyval.ml import RandomSurvivalForest, SurvivalTree
+"""Deprecated alias package.
 
-from .parallel import ParallelModel
-from .series import SeriesModel
+``surpyval.experimental`` was renamed: the alpha-stage models live in
+``surpyval.alpha`` and the survival tree / random survival forest in
+``surpyval.beta.ml``. This module re-exports both for backwards
+compatibility and will be removed in a future release.
+"""
+
+import warnings
+
+from surpyval.alpha import ParallelModel, SeriesModel
+from surpyval.beta.ml import RandomSurvivalForest, SurvivalTree
+
+warnings.warn(
+    "surpyval.experimental is deprecated: use surpyval.alpha "
+    "(ParallelModel, SeriesModel) or surpyval.beta.ml "
+    "(SurvivalTree, RandomSurvivalForest) instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
